backend/requirements.txt
amqp==5.2.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asgiref==3.8.1
asttokens==2.4.1
attrs==23.2.0
billiard==4.2.0
boto3==1.34.95
botocore==1.34.95
celery==5.3.6
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
coreapi==2.3.3
coreschema==0.0.4
cron-descriptor==1.4.3
decorator==5.1.1
Django==5.0.4
django-celery-beat==2.6.0
django-cors-headers==4.3.1
django-cte==1.3.2
django-filter==24.2
django-polymorphic==3.1.0
django-rest-framework==0.1.0
django-storages==1.14.2
django-timezone-field==6.1.0
djangorestframework==3.15.1
drf-spectacular==0.27.2
drf-spectacular-sidecar==2024.4.1
executing==2.0.1
gunicorn==22.0.0
idna==3.7
inflection==0.3.1
iniconfig==2.0.0
ipython==8.24.0
itypes==1.2.0
jedi==0.19.1
Jinja2==3.1.3
jmespath==1.0.1
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
kombu==5.3.6
lazy-object-proxy==1.10.0
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
openapi-codec==1.3.2
packaging==24.0
parso==0.8.4
pexpect==4.9.0
pika==1.3.2
pluggy==1.4.0
prompt-toolkit==3.0.43
psycopg2==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.22
Pygments==2.17.2
pytest==7.4.4
pytest-assert-utils==0.3.1
pytest-common-subject==1.0.6
pytest-django==4.8.0
pytest-drf==1.1.3
pytest-fixture-order==0.1.4
pytest-lambda==1.3.0
python-crontab==3.0.0
python-dateutil==2.9.0.post0
PyYAML==6.0.1
referencing==0.34.0
requests==2.31.0
rpds-py==0.18.0
s3transfer==0.10.1
simplejson==3.19.2
six==1.16.0
sqlparse==0.4.4
stack-data==0.6.3
traitlets==5.14.3
typing_extensions==4.11.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
vine==5.1.0
wcwidth==0.2.13
wrapt==1.16.0
